Fixed: reworked how console output appears so that stderr works
This seems like a much better setup: the React code is all about, and only
about, the interface. There's a new top-level object that deals with rendering
output and giving input to PyPy.js. And that top-level object uses PyPy.js'
repl() method to work around a bug in the usual stderr handling.