Fixed: Verovio renders properly again
When I made the change in how stdout/stderr were rendered, the Verovio
component wasn't rendering properly any more. It's because the SVG notation
got wiped out whenever anything was outputted to the terminal. Now, the
thing that holds the notation is only changed if the "meiForVerovio" prop
isn't empty (that is, only changed if the Verovio component is receiving
new score data).