Fixed: problem with devserver where the CSS file wasn't generated properly
Previously, the "less_autocompile" script would only recompile the LESS file
when it was changed, even if it had never been compiled before (i.e., the
even when the CSS file was missing). Now that script compiles the LESS file
whenever it's started, then it starts watching the file and recompiles on
changes.