WIP: switch out PyPy.js in favour of PyPy-Fujian (tentative)
I'm experimenting with a server-side-on-localhost PyPy runtime for nCoda.
It appears to be faster than PyPy.js, and adopting it for the prototype
would relieve us of a lot of experimental work related to PyPy.js. We're
not yet committed to this approach, but this WIP starts to show that it
would work.