Refactor: Fujian stuff into a dedicated module AND make signals for it
This way, all the details of Fujian interaction are contained in a single
module. Interaction ([re]starting, stopping, submission of input, and acceptance
of output) all takes places via NuclearJS signals. As it happens, most of the
Fujian-related signals don't change a Store directly, but that's not the point.

What this really means is that switching back to PyPy.js, either gradually or
all at once, can easily take place from within a single JS module.