Refactor: "Separator" React component for better resizing
Previously, if you moved your mouse too quickly while using the Separator
to resize CodeScoreView, the Separator wouldn't be able to move so fast,
and "get "lost." Result: it was really difficult to resize anything.

Now, the Separator puts out a <div> over the whole app, and listens for mouse
events on that. This way, even if the Separator doesn't move quickly enough,
it will catch up again.