Change: log.error() and warn() now show modal warning too
Previously, errors and warnings would be written to the console only, just
like info and debug log messages. Because we want users to be alerted when
there is an error (and potentially also when there is a warning) we will
now also show a message when log.error() and warn() are called.