Fixed: copyrightable years in copyright statements
I misunderstood what counts as a copyrightable year, but have now been informed
otherwise by https://www.gnu.org/licenses/gpl-howto.html .

The only copyrightable years are those in which a release is finished that
includes the copyrighted file. We did not make or finish a Julius release in
2015, so this is not a copyrightable year. Therefore I've updated the copyright
statements to show 2016 as the first (and only) copyrightable year.