Change: amazeUI theme is built by importing their LESS stylesheets
Previously, we had our own copies of the amazeUI LESS stylesheets. This would
make it much more difficult to upgrade our amazeUI versions in the future.
Now, our nCoda theme is built by importing and overriding variables in the
amazeUI LESS stylesheets only as needed.

This also eliminates an MIT-licensed file in the Julius repository, making
it easier for external programmers to know the copyright terms of our files.