WIP: Improve integration between amazeUI theme and Julius stylesheets
We should work toward setting everything with amazeUI variables, and
overriding as few things as possible. The goal is to separate generic
style rules that constitute the visual theme from specific style rules
that are required for Julius components to function properly.

This way, upgrading amazeUI, changing to another CSS widget set, and changing
the visual theme, will all be much easier.