Fixed: amazeUI buttons weren't changing properly on :hover
I was hoping to have "outline buttons," where the background is not coloured
but the border and text are. Because of how the amazeUI stylesheets implement
the :hover property, the buttons then don't get darker any more on :hover.
It's not worth my time troubleshooting, so I'll just switch back to whatever.