New: OffCanvas, a generic React component for an off-screen panel
This will be used in the "Staves Structure" part of StructureView, as I
implement Simon's idea for how to deal with the fact that the staves may
be different between sections.

I also ported the nCoda main menu to use OffCanvas. That should have been
in a different commit but I messed up and nobody will call me on it. If
you want to, go ahead, call the police. They're going to arrest me for
using Git badly? Come on, whaddya think this is?!

Part of T2