Remove: unneeded hamburger menu
The amazeUI Topbar React component uses a collapsible menu button when the screen
width is smaller. However, we're not using the built-in collapsible menu, so the
hamburger button is unnecessary. This commit hides the hamburger button by replacing
it with an empty div that is hidden automatically by React.

For T28