Implement: PDFViewer React component to display PDFs with PDF.js
Create new React component, PDFViewer, to display PDFs in Julius.
This will support the display of Lilypond PDFs when that feature is implemented
and is used here to display a dummy PDF on the Lilypond PDF tab in ScoreView.
PDFViewer can be reused to display a PDF anywhere in Julius.

Pages of a rendered PDF are differentiated with shadows and spacing.
Widths of <canvas> elements are set in styles, allowing content to scale
when panes or viewport are resized.

There are buttons to download the PDF and refresh its display.

For T173, T151, and T164