Change: CodeScoreView and ScoreView components for "verovio" Redux store
The approach to using Vida has changed significantly, and it is much more
idiomatic for React. Now, each ScoreView component manages its own VidaView
instance, connected to the global VidaController.

The "verovio" Redux store holds the MEI document itself, rather than a
pseudo-pointer to the VidaView location in an array. This means the MEI
document is held in memory twice (once in the store and once in the VidaView)
but that may never be a problem, and this approach has many advantages too.

For T183