Change: Python/LilyPond code submit btns also save editor
When Python or LilyPond is submitted, Lychee wipes out any saved editor files
to make sure Julius won't get something out-of-date. Because Julius knows that
the editor being submitted is definitely up-to-date, we can save that editor
again to help prevent unnecessary data (formatting) loss.

It's an incomplete solution, but you know...

For T194